chore: I need to use threads ehhhh
