+++ src/server.rs
@@ -1,8 +1,14 @@
-use crate::conn::Connection;
-use crate::Motd;
+#![feature(in_band_lifetimes)]
+use netrex_events::Channel;
+use tokio::net::UdpSocket;
+use tokio::time::sleep;
+
+use crate::conn::{Connection, ConnectionState};
+use crate::{Motd, from_tokenized};
 use std::collections::HashMap;
+use std::net::SocketAddr;
 use std::sync::{Arc, Mutex};
-use std::time::SystemTime;
+use std::time::{SystemTime, Duration};
 
 pub enum RakNetVersion {
     MinecraftRecent,
@@ -21,7 +27,7 @@ impl RakNetVersion {
 }
 
 #[derive(Clone, Debug)]
-pub enum RakNetEvent {
+pub enum RakEvent {
     /// When a connection is created
     ///
     /// ! This is not the same as connecting to the server !
@@ -65,15 +71,15 @@ pub enum RakNetEvent {
     ComplexBinaryError(String, Vec<u8>, String),
 }
 
-impl RakNetEvent {
+impl RakEvent {
     pub fn get_name(&self) -> String {
         match self.clone() {
-            RakNetEvent::ConnectionCreated(_) => "ConnectionCreated".into(),
-            RakNetEvent::Disconnect(_, _) => "Disconnect".into(),
-            RakNetEvent::GamePacket(_, _) => "GamePacket".into(),
-            RakNetEvent::Motd(_, _) => "Motd".into(),
-            RakNetEvent::Error(_) => "Error".into(),
-            RakNetEvent::ComplexBinaryError(_, _, _) => "ComplexBinaryError".into(),
+            RakEvent::ConnectionCreated(_) => "ConnectionCreated".into(),
+            RakEvent::Disconnect(_, _) => "Disconnect".into(),
+            RakEvent::GamePacket(_, _) => "GamePacket".into(),
+            RakEvent::Motd(_, _) => "Motd".into(),
+            RakEvent::Error(_) => "Error".into(),
+            RakEvent::ComplexBinaryError(_, _, _) => "ComplexBinaryError".into(),
         }
     }
 }
@@ -100,9 +106,10 @@ pub enum RakResult {
 pub struct RakNetServer {
     pub address: String,
     pub version: RakNetVersion,
-    pub connections: Arc<Mutex<HashMap<String, Connection>>>,
+    pub connections: Mutex<HashMap<String, Connection>>,
     pub start_time: SystemTime,
-    pub motd: Arc<Motd>,
+    pub motd: Motd,
+    stop: bool,
 }
 
 impl RakNetServer {
@@ -110,14 +117,11 @@ impl RakNetServer {
         Self {
             address,
             version: RakNetVersion::MinecraftRecent,
-            connections: Arc::new(Mutex::new(HashMap::new())),
+            connections: Mutex::new(HashMap::new()),
             start_time: SystemTime::now(),
-            motd: Arc::new(Motd::default()),
+            motd: Motd::default(),
+            stop: false
         }
-    }
-
-    pub fn set_motd(&mut self, motd: Motd) {
-        *Arc::get_mut(&mut self.motd).unwrap() = motd;
     }
 
     /// Sends a stream to the specified address.
@@ -131,5 +135,107 @@ impl RakNetServer {
             Some(c) => c.send(stream, instant),
             None => return,
         };
+    }
+
+    pub async fn start<'a>(&mut self, send_channel: Channel<'a, RakEvent, RakResult>) {
+        let socket = UdpSocket::bind(self.address.parse::<SocketAddr>().expect("Failed to bind to address.")).await.unwrap();
+
+        async {
+            while !&self.stop {
+                if let Err(_) = socket.readable().await {
+                    continue;
+                };
+
+                let mut buf = [0; 2048];
+                if let Ok((len, addr)) = socket.recv_from(&mut buf).await {
+                    let data = &buf[..len];
+
+                    println!("[RakNet] {}: Sent packet: {:?}", addr, &data);
+
+                    if let Ok(mut clients) = self.connections.lock() {
+                        if let Some(c) = clients.get_mut(&addr.ip().to_string()) {
+                            c.recv(&data.to_vec());
+                        } else {
+                            // add the client!
+                            // we need to add cooldown here eventually.
+                            if !self.connections.lock().unwrap().contains_key(&addr.ip().to_string()) {
+                                let mut c = Connection::new(addr, self.start_time, self.motd.clone());
+                                c.recv(&data.to_vec());
+                                clients.insert(addr.ip().to_string(), c);
+                            } else {
+                                // throw an error, this should never happen.
+                                println!("[RakNet] Failed to add client, address already exists but failed mutable borrow.");
+                            }
+                        }
+                    }
+                } else {
+                    // log error in future!
+                    println!("[RakNet] Unknown error decoding packet!");
+                    continue;
+                }
+            }
+        }.await;
+
+        async {
+            while !&self.stop {
+                if let Err(_) = socket.writable().await {
+                    continue;
+                };
+
+                // sleep an entire tick
+                sleep(Duration::from_millis(50)).await;
+
+                let mut clients = self.connections.lock().unwrap();
+                for (addr, _) in clients.clone().iter() {
+                    let client = clients.get_mut(addr).expect("Could not get connection");
+                    client.do_tick();
+
+                    let dispatch = client.event_dispatch.clone();
+                    client.event_dispatch.clear();
+
+                    // emit events if there is a listener for the
+                    for event in dispatch.iter() {
+                        // println!("DEBUG => Dispatching: {:?}", &event.get_name());
+                        if let Some(result) = send_channel.send(event.clone()) {
+                            match result {
+                                RakResult::Motd(_v) => {
+                                    // we don't really support changing
+                                    // client MOTD at the moment...
+                                    // so we don't do anything for this.
+                                }
+                                RakResult::Error(v) => {
+                                    // Calling error forces an error to raise.
+                                    panic!("{}", v);
+                                }
+                                RakResult::Disconnect(_) => {
+                                    client.state = ConnectionState::Offline; // simple hack
+                                    break;
+                                }
+                            }
+                        }
+                    }
+
+                    if client.state == ConnectionState::Offline {
+                        clients.remove(addr);
+                        continue;
+                    }
+
+                    if client.send_queue.len() == 0 {
+                        continue;
+                    }
+
+                    for pk in client.clone().send_queue.into_iter() {
+                        match socket.send_to(&pk[..], &from_tokenized(addr.clone())).await
+                        {
+                            // Add proper handling!
+                            Err(e) => eprintln!("Error Sending Packet [{}]: ", e),
+                            Ok(_) => continue, // println!("\nSent Packet [{}]: {:?}", addr, pk)
+                        }
+                    }
+                    client.send_queue.clear();
+                }
+                drop(clients);
+            }
+        }.await;
     }
 }
