+++ test/src/main.rs
@@ -3,8 +3,8 @@ use rakrs::RakEvent;
 use rakrs::RakNetServer;
 use rakrs::RakResult;
 
-#[test]
-pub fn test_server() {
+#[tokio::main]
+pub async fn main() {
     let mut server = RakNetServer::new(String::from("0.0.0.0:19132"));
     let mut motd = Motd {
         name: "Netrex Raknet".to_owned(),
@@ -33,13 +33,8 @@ pub fn test_server() {
         None
     };
     channel.receive(&mut listener);
-    server.start(channel);
-    
 
     println!("Hi I am running concurrently.");
 
-    loop {
-        let result = server.motd.clone();
-        println!("Motd: {:?}", result);
-    }
+    server.start(channel).await;
 }
